requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
PyMySQL>=1.0